fix CI test,delete support for TLS < v1.2, update README.md,

+++ core/src/test/scala/com/pingcap/tispark/tls/JDBCTLSSuite.scala
@@ -21,14 +21,20 @@ import org.scalatest.Matchers.{be, noException}
 
 class JDBCTLSSuite extends SharedSQLContext {
 
+  val TLSEnable = CheckTLSEnable.isEnableTest()
+
   override def beforeAll(): Unit = {
-    conf.set("enableJDBCSSL", "true")
-    super.beforeAll()
+    if (TLSEnable) {
+      conf.set("enableJDBCSSL", "true")
+      super.beforeAll()
+    }
   }
 
   override def afterAll(): Unit = {
-    tidbStmt.execute(s"DROP DATABASE IF EXISTS `TLS_TEST`;")
-    super.afterAll();
+    if (TLSEnable) {
+      tidbStmt.execute(s"DROP DATABASE IF EXISTS `TLS_TEST`;")
+      super.afterAll();
+    }
   }
 
   test("test JDBC driver connect") {
@@ -37,23 +43,27 @@ class JDBCTLSSuite extends SharedSQLContext {
   }
 
   test("test JDBC connection is SSL") {
-    val result = tidbStmt.executeQuery("SHOW STATUS LIKE \"%Ssl_cipher%\";")
-    while (result.next()) {
-      if (result.getString("Variable_name").equals("Ssl_cipher")) {
-        assert(!result.getString("Value").equals(""))
+    if (TLSEnable) {
+      val result = tidbStmt.executeQuery("SHOW STATUS LIKE \"%Ssl_cipher%\";")
+      while (result.next()) {
+        if (result.getString("Variable_name").equals("Ssl_cipher")) {
+          assert(!result.getString("Value").equals(""))
+        }
       }
     }
   }
 
-  test("test JDBC func") {
-    tidbStmt.execute("CREATE DATABASE IF NOT EXISTS `TLS_TEST`;")
-    tidbStmt.execute(
-      "CREATE TABLE IF NOT EXISTS `TLS_TEST`.`tls_test_table`(id int, name varchar (128)); ")
-    tidbStmt.execute("INSERT INTO `TLS_TEST`.`tls_test_table` VALUES (1, 'jack');")
-    val result = tidbStmt.executeQuery("SELECT * FROM `TLS_TEST`.`tls_test_table`")
-    while (result.next()) {
-      assert(result.getInt("id").equals(1))
-      assert(result.getString("name").equals("jack"))
+  test("test JDBC func with SSL") {
+    if (TLSEnable) {
+      tidbStmt.execute("CREATE DATABASE IF NOT EXISTS `TLS_TEST`;")
+      tidbStmt.execute(
+        "CREATE TABLE IF NOT EXISTS `TLS_TEST`.`tls_test_table`(id int, name varchar (128)); ")
+      tidbStmt.execute("INSERT INTO `TLS_TEST`.`tls_test_table` VALUES (1, 'jack');")
+      val result = tidbStmt.executeQuery("SELECT * FROM `TLS_TEST`.`tls_test_table`")
+      while (result.next()) {
+        assert(result.getInt("id").equals(1))
+        assert(result.getString("name").equals("jack"))
+      }
     }
   }
 }
